test: Added closeToVector utility function (#3)
I wasn't able to use new version of `flame_test`, because it depends on Flutter SDK, and we want Radiance to be a pure-Dart project.

The interim solution is to simply copy the `closeToVector` method over from `flame_test`, and use that. In the future we might be able to use `flame_test` directly.