feat: Added sandbox (#27)
This is an initial implementation of the Sandbox -- eventually it should get many more features, including the ability to create and edit scenes.

The Sandbox is a bit like Examples book in the Flame repository, only here the ultimate goal is to allow the user to modify each scene as they see fit. In particular, this will allow the user to experiment with different settings, different mixtures of behaviors, etc.